+++ app/src/main/java/nsu/ritersport/focususerrs/data/model/PicturesDto.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "pictures")
 data class PicturesDto(
-    @PrimaryKey (autoGenerate = true) val picsId: Int = 0,
+    @PrimaryKey (autoGenerate = true) val picsId: Long = 0,
     val largePicture: String,
     val mediumPicture: String,
     val thumbnailPicture: String,
