+++ app/src/main/java/nsu/ritersport/focususerrs/data/model/UserDto.kt
@@ -32,6 +32,6 @@ data class UserDto(
     val registrationDate: String,
 
     //refs
-    val locationIdRef: Int,
-    val picsRef: Int,
+    val locationIdRef: Long,
+    val picsRef: Long,
 )
